pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lightrag_mcp"
version = "0.1.0"
description = "MCP Server for LightRAG"
requires-python = ">=3.11"
readme = "README.md"
authors = [
    {name = "shemhamforash", email = "killsterak16@gmail.com"},
]
dependencies = [
    "mcp>=1.2.0",
    "httpx>=0.28.1",
    "pydantic>=2.11",
    "python-dotenv>=1.0.1",
    "attrs>=25.3.0",
    "starlette>=0.40.0",
    "uvicorn>=0.32.0",
    "anyio>=4.6.0",
    "fastapi>=0.115.0"
]

[project.optional-dependencies]
dev = [
    "mypy>=1.5.0",
    "ruff>=0.11.4"
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.28.1",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0"
]

[project.scripts]
lightrag-mcp = "lightrag_mcp.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lightrag_mcp"]
